requirements.txt
transformers>=4.40.0
accelerate
einops
timm
gradio
manga-ocr
easyocr
pillow
numpy
protobuf
scipy